backend/requirements.txt
Flask==2.2.5
Flask-Login==0.6.2
Flask-Mail==0.9.1
Flask-SQLAlchemy==3.0.5
Flask-CORS==4.0.0
SQLAlchemy==1.4.49
Werkzeug==2.2.3
python-dotenv==0.21.1
google-generativeai==0.3.2
langchain==0.0.350
langchain-google-genai==0.0.6
langchain-community==0.0.10
faiss-cpu==1.7.4
sentence-transformers==2.2.2
youtube-transcript-api==0.6.1
PyMuPDF==1.23.7
python-docx==0.8.11
reportlab==3.6.13
gunicorn==20.1.0
